+++ examples/js/tweets/close.js
@@ -1,0 +1,4 @@
+	//
+	// Angular bootstrap
+	angular.bootstrap(document, ['twitter-example']);
+});