+++ examples/js/tweets/open.js
@@ -1,0 +1,3 @@
+lux.config({});
+
+require(['angular'], function (angular) {