+++ examples/tweets/config.py
@@ -1,3 +1,7 @@
+REQUIRE_JS = ('https://cdnjs.cloudflare.com/ajax/libs/require.js/2.1.20/'
+              'require.js')
+BOOTSTRAP = 'https://bootswatch.com/flatly/bootstrap.min.css'
+
 twitter_api_key = 'twitter API key of your registered twitter application'
 twitter_api_secret = 'consumer secret'
 twitter_access_token = 'Access token'